# x.json2.decoder2: attributes