```sh
V cache folder /home/hitalo/.vmodules/.cache was wiped.
V tmp.c and tests folder folder /tmp/v_1000 was wiped.
Starting benchmark...
max_iterations: 1000000

***Structure and maps***
 SPENT   347.086 ms in decoder2.decode[Stru](json_data)!
 SPENT   475.593 ms in old_json.decode(Stru, json_data)!

 SPENT   394.047 ms in decoder2.decode[SumTypes](json_data)!
 SPENT   503.372 ms in old_json.decode(SumTypes, json_data)!

 SPENT    88.243 ms in decoder2.decode[StructType[string]](json_data1)!
 SPENT    97.144 ms in old_json.decode(StructType[string], json_data1)!

 SPENT    89.701 ms in decoder2.decode[StructTypeOption[string]](json_data1)!
 SPENT    98.341 ms in old_json.decode(StructTypeOption[string], json_data1)!

 SPENT    69.165 ms in decoder2.decode[StructType[int]](json_data2)!
 SPENT   123.711 ms in old_json.decode(StructType[int], json_data2)!

 SPENT   191.122 ms in decoder2.decode[map[string]string](json_data1)!
 SPENT   198.495 ms in old_json.decode(map[string]string, json_data1)!


***arrays***
 SPENT   331.274 ms in decoder2.decode[[]int]('[1, 2, 3, 4, 5, 6, 7, 8, 9, 0]')!
 SPENT   560.674 ms in old_json.decode([]int, '[1, 2, 3, 4, 5, 6, 7, 8, 9, 0]')!


***simple types***
 SPENT    28.513 ms in decoder2.decode[int]('2')!
 SPENT    26.833 ms in decoder2.decode[bool]('true')!
 SPENT    60.771 ms in decoder2.decode[time.Time]('2022-03-11T13:54:25')!
 SPENT    88.026 ms in decoder2.decode[string]('"abcdefghijklimnopqrstuv"')!

***alias***
 SPENT    27.310 ms in decoder2.decode[IntAlias](2)!
 SPENT    89.452 ms in decoder2.decode[StringAlias]('"abcdefghijklimnopqrstuv"')!

***Sumtypes***
 SPENT    62.700 ms in decoder2.decode[SumTypes](2)!
 SPENT   130.320 ms in decoder2.decode[SumTypes]('"abcdefghijklimnopqrstuv"')!
```