@spytheman 2024-11-03
> Look at this simple solution for dealing with attributes and not having to worry about new attributes that are in fashion.

What is `simple` about it?
Can you expand a bit on that?