@enghitalo 2024-11-03
> > Look at this simple solution for dealing with attributes and not having to worry about new attributes that are in fashion.
> 
> What is `simple` about it?
> Can you expand a bit on that?

Previous solutions had rules, related to attributes, scattered throughout encode_value/decode_value.Now, the rules are distributed across other functions, each with its own responsibility. Which reduces the maintenance "cost". And it leaves more freedom for the user, in case they want to implement a new feature that they think is cool and that has not been developed by the json library. It also allows other libraries to create support for other attributes, such as email and phone numbers validators... ,without much effort and without affecting performance too much.