@spytheman 2024-11-03
The maintenance cost, of code that uses indirection and reflection is higher.
Please provide concrete examples, of code that will be simplified.