@enghitalo 2024-12-02
There is some impact on performance see `***simple types***`
![image](https://github.com/user-attachments/assets/595ab796-c77c-47c4-bed1-32237ac06910)