@enghitalo 2024-12-03
Note: struct is 2x slower to decode when using `@[skip]`
![image](https://github.com/user-attachments/assets/bd45e781-86c4-45b8-9a37-dc090b98a0de)