@enghitalo 2024-12-05
# Before [latest V] (faster)
```sh
V cache folder /home/hitalo/.vmodules/.cache was wiped.
V tmp.c and tests folder folder /tmp/v_1000 was wiped.
Starting benchmark...
max_iterations: 1000000

***Structure and maps***
 SPENT   340.470 ms in decoder2.decode[Stru](json_data)!
 SPENT   482.235 ms in old_json.decode(Stru, json_data)!

 SPENT   387.807 ms in decoder2.decode[SumTypes](json_data)!
 SPENT   509.426 ms in old_json.decode(SumTypes, json_data)!

 SPENT    90.053 ms in decoder2.decode[StructType[string]](json_data1)!
 SPENT    94.422 ms in old_json.decode(StructType[string], json_data1)!

 SPENT    90.643 ms in decoder2.decode[StructTypeOption[string]](json_data1)!
 SPENT    93.809 ms in old_json.decode(StructTypeOption[string], json_data1)!

 SPENT    58.014 ms in decoder2.decode[StructType[int]](json_data2)!
 SPENT   100.996 ms in old_json.decode(StructType[int], json_data2)!

 SPENT   193.633 ms in decoder2.decode[map[string]string](json_data1)!
 SPENT   193.691 ms in old_json.decode(map[string]string, json_data1)!


***arrays***
 SPENT   325.138 ms in decoder2.decode[[]int]('[1, 2, 3, 4, 5, 6, 7, 8, 9, 0]')!
 SPENT   565.424 ms in old_json.decode([]int, '[1, 2, 3, 4, 5, 6, 7, 8, 9, 0]')!


***simple types***
 SPENT    20.355 ms in decoder2.decode[int]('2')!
 SPENT    17.608 ms in decoder2.decode[bool]('true')!
 SPENT    53.068 ms in decoder2.decode[time.Time]('2022-03-11T13:54:25')!
 SPENT    87.921 ms in decoder2.decode[string]('"abcdefghijklimnopqrstuv"')!

***alias***
 SPENT    19.488 ms in decoder2.decode[IntAlias](2)!
 SPENT    89.537 ms in decoder2.decode[StringAlias]('"abcdefghijklimnopqrstuv"')!

***Sumtypes***
 SPENT    73.181 ms in decoder2.decode[SumTypes](2)!
 SPENT   134.789 ms in decoder2.decode[SumTypes]('"abcdefghijklimnopqrstuv"')!
```

Now [This PR] Slower
```sh
V cache folder /home/hitalo/.vmodules/.cache was wiped.
V tmp.c and tests folder folder /tmp/v_1000 was wiped.
Starting benchmark...
max_iterations: 1000000

***Structure and maps***
 SPENT   435.348 ms in decoder2.decode[Stru](json_data)!
 SPENT   481.752 ms in old_json.decode(Stru, json_data)!

 SPENT   502.539 ms in decoder2.decode[SumTypes](json_data)!
 SPENT   502.240 ms in old_json.decode(SumTypes, json_data)!

 SPENT   116.154 ms in decoder2.decode[StructType[string]](json_data1)!
 SPENT    91.710 ms in old_json.decode(StructType[string], json_data1)!

 SPENT   127.106 ms in decoder2.decode[StructTypeOption[string]](json_data1)!
 SPENT    91.615 ms in old_json.decode(StructTypeOption[string], json_data1)!

 SPENT    79.565 ms in decoder2.decode[StructType[int]](json_data2)!
 SPENT    99.632 ms in old_json.decode(StructType[int], json_data2)!

 SPENT   199.808 ms in decoder2.decode[map[string]string](json_data1)!
 SPENT   199.536 ms in old_json.decode(map[string]string, json_data1)!


***arrays***
 SPENT   331.909 ms in decoder2.decode[[]int]('[1, 2, 3, 4, 5, 6, 7, 8, 9, 0]')!
 SPENT   571.260 ms in old_json.decode([]int, '[1, 2, 3, 4, 5, 6, 7, 8, 9, 0]')!


***simple types***
 SPENT    19.737 ms in decoder2.decode[int]('2')!
 SPENT    17.488 ms in decoder2.decode[bool]('true')!
 SPENT    53.500 ms in decoder2.decode[time.Time]('2022-03-11T13:54:25')!
 SPENT    82.150 ms in decoder2.decode[string]('"abcdefghijklimnopqrstuv"')!

***alias***
 SPENT    19.824 ms in decoder2.decode[IntAlias](2)!
 SPENT    80.840 ms in decoder2.decode[StringAlias]('"abcdefghijklimnopqrstuv"')!

***Sumtypes***
 SPENT    63.274 ms in decoder2.decode[SumTypes](2)!
 SPENT   130.163 ms in decoder2.decode[SumTypes]('"abcdefghijklimnopqrstuv"')!
```