@enghitalo 2024-12-05
After merge on latest master

```sh
V cache folder /home/hitalo/.vmodules/.cache was wiped.
V tmp.c and tests folder folder /tmp/v_1000 was wiped.
Starting benchmark...
max_iterations: 1000000

***Structure and maps***
 SPENT   404.889 ms in decoder2.decode[Stru](json_data)!
 SPENT   460.689 ms in old_json.decode(Stru, json_data)!

 SPENT   479.434 ms in decoder2.decode[SumTypes](json_data)!
 SPENT   494.619 ms in old_json.decode(SumTypes, json_data)!

 SPENT   110.415 ms in decoder2.decode[StructType[string]](json_data1)!
 SPENT    89.593 ms in old_json.decode(StructType[string], json_data1)!

 SPENT   119.058 ms in decoder2.decode[StructTypeOption[string]](json_data1)!
 SPENT    87.923 ms in old_json.decode(StructTypeOption[string], json_data1)!

 SPENT    73.912 ms in decoder2.decode[StructType[int]](json_data2)!
 SPENT    94.753 ms in old_json.decode(StructType[int], json_data2)!

 SPENT   183.044 ms in decoder2.decode[map[string]string](json_data1)!
 SPENT   183.367 ms in old_json.decode(map[string]string, json_data1)!


***arrays***
 SPENT   311.392 ms in decoder2.decode[[]int]('[1, 2, 3, 4, 5, 6, 7, 8, 9, 0]')!
 SPENT   553.583 ms in old_json.decode([]int, '[1, 2, 3, 4, 5, 6, 7, 8, 9, 0]')!


***simple types***
 SPENT    19.128 ms in decoder2.decode[int]('2')!
 SPENT    17.133 ms in decoder2.decode[bool]('true')!
 SPENT    52.787 ms in decoder2.decode[time.Time]('2022-03-11T13:54:25')!
 SPENT    84.898 ms in decoder2.decode[string]('"abcdefghijklimnopqrstuv"')!

***alias***
 SPENT    19.005 ms in decoder2.decode[IntAlias](2)!
 SPENT    83.076 ms in decoder2.decode[StringAlias]('"abcdefghijklimnopqrstuv"')!

***Sumtypes***
 SPENT    59.625 ms in decoder2.decode[SumTypes](2)!
 SPENT   125.958 ms in decoder2.decode[SumTypes]('"abcdefghijklimnopqrstuv"')!
```